Create GraphicalSequenceAlgorithm.py
This script implements the Graphical Sequence Algorithm, which determines whether a sequence of degrees for a graph G is graphical.

This was implemented in Python. The decision to do this algorithm in Python was mainly due to a want to learn and get used to the Python language and syntax.